refactor: ItemNameCell -> sep component for LinkPopup
